refactor(validator): reorganize rental routes into routes module
* Move rental_routes.rs into routes/rentals.rs for better organization
* Update imports to use routes::* instead of rental_routes::*
* Add rentals module export to routes/mod.rs
* Update external imports in basilica-api, sdk, and cli modules
* Maintain consistent module structure with other route handlers
