+++ crates/basilica-validator/src/api/routes/mod.rs
@@ -5,6 +5,7 @@ pub mod config;
 pub mod gpu;
 pub mod health;
 pub mod miners;
+pub mod rentals;
 pub mod verification;
 
 pub use capacity::*;
@@ -12,4 +13,5 @@ pub use config::*;
 pub use gpu::*;
 pub use health::*;
 pub use miners::*;
+pub use rentals::*;
 pub use verification::*;
