+++ crates/basilica-validator/src/api/routes/rentals.rs
